Requests are done
removed the onsolution back as that was merely outputting whatever solutoin was found - it may not have been the best solution (i.e. the solution that meets all teh requests).

Thus that call back is removed and the assertions are done once the solve method returns.

Thus the solver is returned from the solve method so that the calling context can access the values of the variables.